requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=9.4.0
tensorboard>=2.12.0
jupyter>=1.0.0
scikit-learn>=1.2.0
scipy>=1.10.0
tqdm>=4.64.0
opencv-python>=4.7.0
pyyaml>=6.0